% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{adf.testWindow}
\alias{adf.testWindow}
\title{tseries adf.test Window (adf.testWindow)}
\usage{
adf.testWindow(data, stepSize, windowSize)
}
\arguments{
\item{data}{The time series data}

\item{stepSize}{The about of elements to jump after each window}

\item{windowSize}{The size of each window being sampled}
}
\value{
A vector of the computed values for each window returned from the adf.test function
}
\description{
Computes the stationarityy using tseries::adf.test of several windows of the original data
}

