% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeRUL}
\alias{computeRUL}
\title{Compute Remaining Useful Life (computeRUL)}
\usage{
computeRUL(lifeTab, accDegVal)
}
\arguments{
\item{lifeTab}{The life table created from createLifeTab or updateLifeTab.}

\item{accDegVal}{The final accumulated degredation value of your data.}
}
\value{
The estimated remaining useful life of the system.
}
\description{
Determines the Remaining useful life of a system based on its accumulated degradation value
}

