% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert}
\alias{convert}
\title{DMarkov Convert (convert)}
\usage{
convert(data, groupSize, alphabetSize)
}
\arguments{
\item{data}{A vector containing the SAX series to be converted}

\item{groupSize}{The chosen group size for sampling from original series}

\item{alphabetSize}{The original alphabet size of the inputted SAX data}
}
\value{
A vector containing the new SAX series of alphabet size 'alphabetSize'^'groupSize'
}
\description{
Convert a SAX series to another SAX series with chosen group size.
}

