% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosMeasure}
\alias{cosMeasure}
\title{Causality Measurement (cosMeasure)}
\usage{
cosMeasure(tabTrained, testObs, testHid, testOSize, testHSize)
}
\arguments{
\item{tabTrained}{A NumericMatrix returned from the 'count' function for the "training" data}

\item{testObs}{A vector containing the test observed data}

\item{testHid}{A vector containing the test hidden data}

\item{testOSize}{The alphabet size/cardinality of the observed data}

\item{testHSize}{The alphabet size/cardinality of the hidden data}
}
\description{
A causality measurement between two sets of data
}

