% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count}
\alias{count}
\title{Count Occurrences (count)}
\usage{
count(dataO, dataH, oSize, hSize)
}
\arguments{
\item{dataO}{A vector containing observed data}

\item{dataH}{A vector containing hidden data}

\item{oSize}{The alphabet size/cardinality of the observed data}

\item{hSize}{The alphabet size/cardinality of the hidden data}
}
\value{
A NumericMatrix of times each combination of Observed followed by Hidden has occured
}
\description{
Count the number of times an Observed state 'm' is followed by a Hidden Symbol 'n'
}

