% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createLifeTab}
\alias{createLifeTab}
\title{Create Life Table (createLifeTab)}
\usage{
createLifeTab(accDeg, timeInterval)
}
\arguments{
\item{accDeg}{The output of the 'accDegrad' function. A vector containing the Accumulated Degredation values of the time series data.}

\item{timeInterval}{The interval of time inbetween each sample of data.}
}
\value{
A Life Table used for RUL computation
}
\description{
Create a Life Table comparing the accumulated Degradation to the time remaining
}

