% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULHMM.R
\name{createModels.path}
\alias{createModels.path}
\title{Create HMM (createModel)}
\usage{
createModels.path(data.path, SAXalphabetSize, SAXgroupSize, exp2,
  hidAlphabetSize)
}
\arguments{
\item{data.path}{The file path to the sets data.}

\item{SAXalphabetSize}{The alphabet size used for the SAX conversion.}

\item{SAXgroupSize}{The size of each group to be converted to a single SAX symbol.}

\item{exp2}{Two to the power of exp2 will be the size.}

\item{hidAlphabetSize}{The alphabet size of the hidden sequence.}
}
\value{
A HMM generated from the time series.
}
\description{
Create a several hidden markov model for the sets of data located in the given file path.
}

