% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createProbMatX}
\alias{createProbMatX}
\title{Create Probability Matrix X (createProbMatX)}
\usage{
createProbMatX(dataO, dataH, groupSize, alphabetSizeO, alphabetSizeH)
}
\arguments{
\item{dataO}{A vector containing the observed data in SAX}

\item{dataH}{A vector containing the hidden data in SAX}

\item{groupSize}{The chosen group size for DMarkov}

\item{alphabetSizeO}{The original alphabet size of the observed data}

\item{alphabetSizeH}{The original alphabet size of the hidden data}
}
\value{
A ProbMatX object. A list containing the transition, emission and reverse emission probabilities trained from the input data
}
\description{
Create new ProbMatX object from a set of observed and hidden data. This ProbMatX object contains transition, emission and reverse emission probabilities
}

