% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT outliers.R
\name{detectOutliersFft}
\alias{detectOutliersFft}
\title{Detect outliers FFT (detectOutliersFft)}
\usage{
detectOutliersFft(data, stepSize, thresholdFreq = 0.1, frequencyAmp = 0.01)
}
\arguments{
\item{data}{A vector containing the time series data}

\item{stepSize}{The size of half a window that is sampled from the data}

\item{thresholdFreq}{Threshold frequency compared with frequency response of the signal to determine outliers}

\item{frequencyAmp}{Frequency Amplitude}
}
\value{
A vector with the indices of the detected outliers
}
\description{
Detecting outliers in a series using FFT(Fast Fourier Transform)
}

