% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distribution}
\alias{distribution}
\title{Compute Distribution (distribution)}
\usage{
distribution(tabTrained, testObs, testHid, testOSize, testHSize, step,
  subSeqSize)
}
\arguments{
\item{tabTrained}{Trained count matrix created from 'count' function}

\item{testObs}{A vector containing the test observed data}

\item{testHid}{A vector containing the test hidden data}

\item{testOSize}{The alphabet size of the observed data}

\item{testHSize}{The alphabet size of the hidden data}

\item{step}{The step size for subsequence sampling}

\item{subSeqSize}{The size of each subsequence extracted from test sequence}
}
\value{
A vector containing the causality measurement (Lambda) of the several small subsequences sampled
}
\description{
Compute the causality measurement between several subsequences and a trained set of data
}

