% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E}
\alias{E}
\title{Energy (E)}
\usage{
E(v, h, a, b, w)
}
\arguments{
\item{v}{A vector containing the binary visible data}

\item{h}{A vector containing the binary hidden data}

\item{a}{A vector containing the bias values for the visible data}

\item{b}{A vector containing the bias values for the hidden data}

\item{w}{A matrix containing the weights between the visible and hidden data}
}
\value{
A double value. The energy of the binary vector of visible data and binary vector of hidden data
}
\description{
Compute energy of a binary vector of visible data and a binary vector of hidden data
}

