% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F}
\alias{F}
\title{Free Energy (F)}
\usage{
F(v, hSize, a, b, w)
}
\arguments{
\item{v}{A vector containing the binary visible data}

\item{hSize}{The size of vector containing the hidden data}

\item{a}{A vector containing the bias values for the visible data}

\item{b}{A vector containing the bias values for the hidden data}

\item{w}{A matrix containing the weights between the visible and hidden data}
}
\value{
A double value. The free energy for that vector of visible data
}
\description{
Compute free energy of a binary vector of visible data for all possible hidden data of chosen size
}

