% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward}
\alias{forward}
\title{Forward Algorithm (forward)}
\usage{
forward(transProb, emisProb, initProb, dataO, dataH, indexH)
}
\arguments{
\item{transProb}{A matrix containing the transition probabilites from the hidden markov model}

\item{emisProb}{A matrix containing the emission probabilities from the hidden markov model}

\item{initProb}{A matrix containing the initial probabilities from the hidden markov model}

\item{dataO}{A vector containing the observed sequence of data}

\item{dataH}{A single symbol of the hidden data}

\item{indexH}{The index at which the single symbol of hidden data above is located in it's original sequence}
}
\value{
A double(real) value. The probability that the hidden symbol is at 'indexH' given the observed sequence and probabilitiy matrices
}
\description{
Forward algorithm for hidden markov models.
}

