% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULHMM.R
\name{meanStdDevR}
\alias{meanStdDevR}
\title{Compute mean and standard deviation (meanStdDevR)}
\usage{
meanStdDevR(hidSeq, hidAlphabetSize)
}
\arguments{
\item{hidSeq}{The hidden sequence. Output of viterbiR.}

\item{hidAlphabetSize}{The alphabet size of the hidden sequence}
}
\value{
Vectors of the mean and median values.
}
\description{
Compute mean and standard deviation of the time in each state
}

