% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{ndiffsWindow}
\alias{ndiffsWindow}
\title{forecast ndiffs Window (ndiffsWindow)}
\usage{
ndiffsWindow(data, stepSize, windowSize)
}
\arguments{
\item{data}{The time series data}

\item{stepSize}{The about of elements to jump after each window}

\item{windowSize}{The size of each window being sampled}
}
\value{
A vector of the computed values for each window returned from the ndiffs function
}
\description{
Computes the stationarityy using forecast::ndiffs(test = "adf") of several windows of the original data
}

