% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULHMM.R
\name{RUL.HMM}
\alias{RUL.HMM}
\title{Remaining useful life using HMM (RUL.HMM)}
\usage{
RUL.HMM(timeSeries, models, SAXalphabetSize, SAXgroupSize, exp2,
  lastStateScanNum, confidenceCoef)
}
\arguments{
\item{timeSeries}{The time series to be converted.}

\item{models}{The generated hidden markov models. Output from createModel(s).*}

\item{SAXalphabetSize}{The alphabet size used for the SAX conversion.}

\item{SAXgroupSize}{The size of each group to be converted to a single SAX symbol.}

\item{exp2}{Two to the power of exp2 will be the size.}

\item{lastStateScanNum}{The amount of states scanned when finding the last sequence state.}

\item{confidenceCoef}{A constant used when computing the lower and upper bounds of the RUL.}
}
\value{
The estimated pessimistic remaining useful life with an upper and lower bound.
}
\description{
Compute pessimistic remaining useful life using HMMs
}

