% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKLD.R
\name{runKLDVec}
\alias{runKLDVec}
\title{KL Divergence of a single time series (runKLDVec)}
\usage{
runKLDVec(rawData, saxGroupSize, saxAlphabetSize, dMarkovSize, disStep,
  disWindowSize, k)
}
\arguments{
\item{rawData}{A vector of the raw data/time series data}

\item{saxGroupSize}{The group size when converting data into SAX representation}

\item{saxAlphabetSize}{The alphabet size when converting data into SAX representation}

\item{dMarkovSize}{The size when grouping observed data to predict the next hidden state}

\item{disStep}{The step size of sampling when computing the distribution}

\item{disWindowSize}{The size of the sample when computing the distribution}

\item{k}{The maximum number of nearest neighbours when computing the KL divergence}
}
\value{
This function does not return anything. However it does create a png of the KL divergence
}
\description{
Convert time series data to SAX, group Train and Test sequences, compute each distribution then graph the KL divergence between the two sequences
}

