% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKLD.R
\name{runKLDVecRand}
\alias{runKLDVecRand}
\title{KL Divergence of a single time series and a random data set (runKLDVecRand)}
\usage{
runKLDVecRand(rawData, saxGroupSize, saxAlphabetSize, delay, dMarkovSize,
  disStep, disWindowSize, k)
}
\arguments{
\item{rawData}{A vector of the raw data/time series data}

\item{saxGroupSize}{The group size when converting data into SAX representation}

\item{saxAlphabetSize}{The alphabet size when converting data into SAX representation}

\item{delay}{The number of zeros to pad on the hidden sequence before converting to SAX in order to delay it}

\item{dMarkovSize}{The size when grouping observed data to predict the next hidden state}

\item{disStep}{The step size of sampling when computing the distribution}

\item{disWindowSize}{The size of the sample when computing the distribution}

\item{k}{The maximum number of nearest neighbours when computing the KL divergence}
}
\value{
This function does not return anything. However it does plot a graph of the KLD
}
\description{
Convert time series data to SAX, randomize test sequence, compute each distribution then graph the KL divergence between the two sequences
}

