% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateHid}
\alias{simulateHid}
\title{Simulate Hidden Series (simulateHid)}
\usage{
simulateHid(probMatX, dataObs, groupSize, alphabetSizeO, alphabetSizeH)
}
\arguments{
\item{probMatX}{The ProbMatX object created from 'createProbMatX' function}

\item{dataObs}{The observed data used to simulate the hidden data}

\item{groupSize}{The groupSize of the DMarkov model. Should be the same as the one used to create ProbMatX object}

\item{alphabetSizeO}{The original alphabet size of the observed data. The alphabetSize must be the same as the one used when creating the ProbMatX object}

\item{alphabetSizeH}{The original alphabet size of the hidden data. The alphabetSize must be the same as the one used when creating the ProbMatX object}
}
\value{
A vector containing the simulated hidden data
}
\description{
Simulate hidden series given an observed series and using the reverse emission probability from ProbMatX object
}

