% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULHMM.R
\name{toWPDSAX}
\alias{toWPDSAX}
\title{To WPD, to SAX (toWPDSAX)}
\usage{
toWPDSAX(timeSeries, SAXalphabetSize, SAXgroupSize, exp2)
}
\arguments{
\item{timeSeries}{The time series to be converted.}

\item{SAXalphabetSize}{The alphabet size used for the SAX conversion.}

\item{SAXgroupSize}{The size of each group to be converted to a single SAX symbol.}

\item{exp2}{Two to the power of exp2 will be the size.}
}
\value{
The SAX sequence of the given time series.
}
\description{
Extract features using Wavelet Packlet Decomposition, compute nodal energies then convert to SAX.
}

