% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{train}
\alias{train}
\title{Train (train)}
\usage{
train(transProb, emisProb, initProb, dataV)
}
\arguments{
\item{transProb}{A matrix containing the transition probabilites from the hidden markov model}

\item{emisProb}{A matrix containing the emission probabilities from the hidden markov model}

\item{initProb}{A matrix containing the initial probabilities from the hidden markov model}

\item{dataV}{A vector containing the observed sequence of data}
}
\value{
A list of new probability matrices
}
\description{
Train Hidden Markov model
}

