% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateProbMatX}
\alias{updateProbMatX}
\title{Update Probability Matrix X (updateProbMatX)}
\usage{
updateProbMatX(prevData, dataO, dataH, groupSize, alphabetSizeO, alphabetSizeH)
}
\arguments{
\item{prevData}{The ProbMatX object created from 'createProbMatX' function}

\item{dataO}{the new observed data to be updated into the ProbMatX object}

\item{dataH}{the new hidden data to be updated into the ProbMatX object}

\item{groupSize}{The chosen group size for DMarkov. The groupSize must be the same as the one used when creating the ProbMatX object}

\item{alphabetSizeO}{The original alphabet size of the observed data. The alphabetSize must be the same as the one used when creating the ProbMatX object}

\item{alphabetSizeH}{The original alphabet size of the hidden data. The alphabetSize must be the same as the one used when creating the ProbMatX object}
}
\value{
A ProbMatX object. A list containing the updated transition, emission and reverse emission probabilities from the input data
}
\description{
Update ProbMatX object with more data
}

