% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{viterbi}
\alias{viterbi}
\title{Determine the most probable hidden sequence given an observed sequence}
\usage{
viterbi(transProb, emisProb, initProb, dataV)
}
\arguments{
\item{transProb}{A matrix containing the transition probabilites from the hidden markov model}

\item{emisProb}{A matrix containing the emission probabilities from the hidden markov model}

\item{initProb}{A matrix containing the initial probabilities from the hidden markov model}

\item{dataV}{A vector containing the observed sequence of data}
}
\value{
A vector containing the most probable sequence of hidden states
}
\description{
Determine the most probable hidden sequence given an observed sequence
}

