% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{viterbiProbVal}
\alias{viterbiProbVal}
\title{Viterbi Probability Value (viterbiProbVal)}
\usage{
viterbiProbVal(transProb, emisProb, initProb, dataO, dataH, logLand = FALSE)
}
\arguments{
\item{transProb}{A matrix containing the transition probabilites from the hidden markov model}

\item{emisProb}{A matrix containing the emission probabilities from the hidden markov model}

\item{initProb}{A matrix containing the initial probabilities from the hidden markov model}

\item{dataO}{A vector containing the observed sequence of data}

\item{dataH}{A vector containing the hidden sequence of data}
}
\value{
A vector containing the probability to get to each index of the hidden state given the observed state
}
\description{
Compute the probability of that the given observed and hidden sequences occur in a trained system
}

