% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULHMM.R
\name{viterbiR}
\alias{viterbiR}
\title{Viterbi Algorithm (viterbiR)}
\usage{
viterbiR(ObsSeq, probMat)
}
\arguments{
\item{ObsSeq}{A discrete observed time series. Output from toWPDSAX.}

\item{probMat}{The probability matrices of the hidden markov model. Output from trainHMM.}
}
\value{
The most probably hidden sequence.
}
\description{
Compute the most probable hidden sequence given an observed sequence and an HMM.
}

