% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalDecomp.R
\name{decomp}
\alias{decomp}
\title{Signal Decomposition (decomp)}
\usage{
decomp(data, peakThreshold, gridColNum)
}
\arguments{
\item{data}{The time series signal as a vector}

\item{peakThreshold}{The chosen threshold for which to identify peaks in the spectrum}

\item{gridColNum}{The number of columns when the displaying the subseries plots. The number of rows is also calculated from this value.}
}
\value{
None. Computes and plots the decompsed subseries of a time series signal
}
\description{
Decompose a time series signal into several subseries using band-pass filtering based decomposition
}

